+++ internal/compose/documentLink_test.go
@@ -1362,6 +1362,219 @@ services:
 		})
 	}
 }
+func TestDocumentLink_ServiceLabelFileLinks(t *testing.T) {
+	testsFolder := filepath.Join(os.TempDir(), t.Name())
+	composeStringURI := fmt.Sprintf("file:///%v", strings.TrimPrefix(filepath.ToSlash(filepath.Join(testsFolder, "compose.yaml")), "/"))
+
+	testCases := []struct {
+		name      string
+		content   string
+		path      string
+		linkRange protocol.Range
+	}{
+		{
+			name: "string value app.labels",
+			content: `
+services:
+  test:
+    label_file: app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 16},
+				End:   protocol.Position{Line: 3, Character: 26},
+			},
+		},
+		{
+			name: "string value ./app.labels",
+			content: `
+services:
+  test:
+    label_file: ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 16},
+				End:   protocol.Position{Line: 3, Character: 28},
+			},
+		},
+		{
+			name: "quoted string value \"./app.labels\"",
+			content: `
+services:
+  test:
+    label_file: "./app.labels"`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 17},
+				End:   protocol.Position{Line: 3, Character: 29},
+			},
+		},
+		{
+			name: "attribute value is null",
+			content: `
+services:
+  test:
+    label_file: null`,
+		},
+		{
+			name: "array items",
+			content: `
+services:
+  test:
+    label_file:
+      - ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 4, Character: 8},
+				End:   protocol.Position{Line: 4, Character: 20},
+			},
+		},
+		{
+			name: "array item is null",
+			content: `
+services:
+  test:
+    label_file:
+      - null`,
+		},
+		{
+			name: "anchors and aliases to nothing",
+			content: `
+services:
+  test:
+    label_file: &anchor
+  test2:
+    label_file: *anchor`,
+		},
+		{
+			name: "anchor on the services object itself",
+			content: `
+&anchor services:
+  test:
+    label_file: ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 16},
+				End:   protocol.Position{Line: 3, Character: 28},
+			},
+		},
+		{
+			name: "anchor on the services object's value",
+			content: `
+services: &anchor
+  test:
+    label_file: ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 16},
+				End:   protocol.Position{Line: 3, Character: 28},
+			},
+		},
+		{
+			name: "anchor on the service object itself",
+			content: `
+services:
+  &anchor test:
+    label_file: ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 16},
+				End:   protocol.Position{Line: 3, Character: 28},
+			},
+		},
+		{
+			name: "anchor on the service object's value",
+			content: `
+services:
+  test: &anchor
+    label_file: ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 16},
+				End:   protocol.Position{Line: 3, Character: 28},
+			},
+		},
+		{
+			name: "anchor on the service object's value as JSON",
+			content: `
+services:
+  test: &anchor { label_file: ./app.labels }`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 2, Character: 30},
+				End:   protocol.Position{Line: 2, Character: 42},
+			},
+		},
+		{
+			name: "anchor on the label_file string attribute itself",
+			content: `
+services:
+  test: &anchor
+    &anchor label_file: ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 24},
+				End:   protocol.Position{Line: 3, Character: 36},
+			},
+		},
+		{
+			name: "anchor on the label_file string attribute's value",
+			content: `
+services:
+  test: &anchor
+    label_file: &anchor ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 3, Character: 24},
+				End:   protocol.Position{Line: 3, Character: 36},
+			},
+		},
+		{
+			name: "anchor on the label_file array attribute's value",
+			content: `
+services:
+  test: &anchor
+    label_file: &anchor
+      - ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 4, Character: 8},
+				End:   protocol.Position{Line: 4, Character: 20},
+			},
+		},
+		{
+			name: "anchor on the label_file array item's value",
+			content: `
+services:
+  test: &anchor
+    label_file:
+      - &anchor ./app.labels`,
+			path: filepath.Join(testsFolder, "./app.labels"),
+			linkRange: protocol.Range{
+				Start: protocol.Position{Line: 4, Character: 16},
+				End:   protocol.Position{Line: 4, Character: 28},
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			mgr := document.NewDocumentManager()
+			doc := document.NewComposeDocument(mgr, "compose.yaml", 1, []byte(tc.content))
+			links, err := DocumentLink(context.Background(), composeStringURI, doc)
+			require.NoError(t, err)
+			if tc.path == "" {
+				require.Equal(t, []protocol.DocumentLink{}, links)
+			} else {
+				link := protocol.DocumentLink{
+					Range:   tc.linkRange,
+					Target:  types.CreateStringPointer(fmt.Sprintf("file:///%v", strings.TrimPrefix(filepath.ToSlash(tc.path), "/"))),
+					Tooltip: types.CreateStringPointer(tc.path),
+				}
+				require.Equal(t, []protocol.DocumentLink{link}, links)
+			}
+		})
+	}
+}
 
 func TestDocumentLink_ConfigFileLinks(t *testing.T) {
 	testsFolder := filepath.Join(os.TempDir(), t.Name())